Improved Map
removed SFTiles in favor of s_tileset. s_tileset holds only the unique
tiles that make up the map and m_data just references which tile to use
from s_tileset for each square on the map.  s_tileset is only generated
once, when the first map is created.  In the future if we want to be
able to load multiple tilesets we will have to change it from being
static variable to a member variable.

I also changed the naming convention to better match the rest of the
code.  Such as using "m_."